+++ packages/core/src/vite-env.d.ts
@@ -1,1 +1,7 @@
 /// <reference types="vite/client" />
+
+declare const process: {
+  env: {
+    NODE_ENV: string;
+  };
+};