+++ packages/mantine/src/BlockNoteView.tsx
@@ -11,7 +11,7 @@ import {
   usePrefersColorScheme,
 } from "@blocknote/react";
 import { MantineContext, MantineProvider } from "@mantine/core";
-import React, { useCallback, useContext } from "react";
+import React, { useCallback, useContext, useEffect } from "react";
 import {
   applyBlockNoteCSSVariablesFromTheme,
   removeBlockNoteCSSVariables,
@@ -38,17 +38,23 @@ export const BlockNoteView = <
         };
   },
 ) => {
-  const { className, theme, ...rest } = props;
+  const { className, theme, editor, ...rest } = props;
 
   const existingContext = useBlockNoteContext();
   const systemColorScheme = usePrefersColorScheme();
   const defaultColorScheme =
     existingContext?.colorSchemePreference || systemColorScheme;
 
-  const ref = useCallback(
-    (node: HTMLDivElement | null) => {
+  const finalTheme =
+    typeof theme === "string"
+      ? theme
+      : defaultColorScheme !== "no-preference"
+        ? defaultColorScheme
+        : "light";
+
+  const applyThemeVariables = useCallback(
+    (node: HTMLElement | null) => {
       if (!node) {
-        // todo: clean variables?
         return;
       }
 
@@ -70,14 +76,15 @@ export const BlockNoteView = <
     [defaultColorScheme, theme],
   );
 
-  const mantineContext = useContext(MantineContext);
+  useEffect(() => {
+    if (!editor.portalElement) {
+      throw new Error("Portal element not found");
+    }
+    editor.portalElement.setAttribute("data-mantine-color-scheme", finalTheme);
+    applyThemeVariables(editor.portalElement);
+  }, [editor, applyThemeVariables, finalTheme]);
 
-  const finalTheme =
-    typeof theme === "string"
-      ? theme
-      : defaultColorScheme !== "no-preference"
-        ? defaultColorScheme
-        : "light";
+  const mantineContext = useContext(MantineContext);
 
   const view = (
     <ComponentsContext.Provider value={components}>
@@ -85,8 +92,9 @@ export const BlockNoteView = <
         data-mantine-color-scheme={finalTheme}
         className={mergeCSSClasses("bn-mantine", className || "")}
         theme={typeof theme === "object" ? undefined : theme}
+        editor={editor}
         {...rest}
-        ref={ref}
+        ref={applyThemeVariables}
       />
     </ComponentsContext.Provider>
   );